Refactor metadata, move common functions to MetaDataParser
