+++ airsonic-main/src/main/java/org/airsonic/player/service/metadata/MetaDataParser.java
@@ -24,9 +24,14 @@ import org.airsonic.player.domain.MusicFolder;
 import org.airsonic.player.service.SettingsService;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang.StringUtils;
+import org.jaudiotagger.tag.reference.GenreTypes;
 
 import java.io.File;
 import java.util.List;
+import java.util.SortedSet;
+import java.util.TreeSet;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 
 /**
@@ -35,6 +40,10 @@ import java.util.List;
  * @author Sindre Mehus
  */
 public abstract class MetaDataParser {
+
+    private static final Pattern GENRE_PATTERN = Pattern.compile("\\((\\d+)\\).*");
+    private static final Pattern TRACK_NUMBER_PATTERN = Pattern.compile("(\\d+)/\\d+");
+    private static final Pattern YEAR_NUMBER_PATTERN = Pattern.compile("(\\d{4}).*");
 
     /**
      * Parses meta data for the given file.
@@ -107,7 +116,7 @@ public abstract class MetaDataParser {
     /**
      * Guesses the artist for the given file.
      */
-    String guessArtist(File file) {
+    public String guessArtist(File file) {
         File parent = file.getParentFile();
         if (isRoot(parent)) {
             return null;
@@ -119,7 +128,7 @@ public abstract class MetaDataParser {
     /**
      * Guesses the album for the given file.
      */
-    String guessAlbum(File file, String artist) {
+    public String guessAlbum(File file, String artist) {
         File parent = file.getParentFile();
         String album = isRoot(parent) ? null : parent.getName();
         if (artist != null && album != null) {
@@ -144,6 +153,103 @@ public abstract class MetaDataParser {
             }
         }
         return false;
+    }
+
+    /**
+     * Returns all tags supported by id3v1.
+     */
+    public static SortedSet<String> getID3V1Genres() {
+        return new TreeSet<String>(GenreTypes.getInstanceOf().getAlphabeticalValueList());
+    }
+
+    /**
+     * Sometimes the genre is returned as "(17)" or "(17)Rock", instead of "Rock".  This method
+     * maps the genre ID to the corresponding text.
+     */
+    protected String mapGenre(String genre) {
+        if (genre == null) {
+            return null;
+        }
+        Matcher matcher = GENRE_PATTERN.matcher(genre);
+        if (matcher.matches()) {
+            int genreId = Integer.parseInt(matcher.group(1));
+            if (genreId >= 0 && genreId < GenreTypes.getInstanceOf().getSize()) {
+                return GenreTypes.getInstanceOf().getValueForId(genreId);
+            }
+        }
+        return genre;
+    }
+
+    /**
+     * Parses the track/disc number from the given string.  Also supports
+     * numbers on the form "4/12".
+     */
+    protected Integer parseTrackNumber(String trackNumber) {
+        if (trackNumber == null) {
+            return null;
+        }
+
+        Integer result = null;
+
+        try {
+            result = Integer.valueOf(trackNumber);
+        } catch (NumberFormatException x) {
+            Matcher matcher = TRACK_NUMBER_PATTERN.matcher(trackNumber);
+            if (matcher.matches()) {
+                try {
+                    result = Integer.valueOf(matcher.group(1));
+                } catch (NumberFormatException e) {
+                    return null;
+                }
+            }
+        }
+
+        if (Integer.valueOf(0).equals(result)) {
+            return null;
+        }
+        return result;
+    }
+
+    protected Integer parseYear(String year) {
+        if (year == null) {
+            return null;
+        }
+
+        Integer result = null;
+
+        try {
+            result = Integer.valueOf(year);
+        } catch (NumberFormatException x) {
+            Matcher matcher = YEAR_NUMBER_PATTERN.matcher(year);
+            if (matcher.matches()) {
+                try {
+                    result = Integer.valueOf(matcher.group(1));
+                } catch (NumberFormatException e) {
+                    return null;
+                }
+            }
+        }
+
+        if (Integer.valueOf(0).equals(result)) {
+            return null;
+        }
+        return result;
+    }
+
+    protected Integer parseInteger(String s) {
+        s = StringUtils.trimToNull(s);
+        if (s == null) {
+            return null;
+        }
+        try {
+            Integer result = Integer.valueOf(s);
+            if (Integer.valueOf(0).equals(result)) {
+                return null;
+            }
+            return result;
+        } catch (NumberFormatException x) {
+            return null;
+        }
     }
 
     abstract SettingsService getSettingsService();
